backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy

redis
pydantic
python-dotenv
numpy
pandas
scikit-learn
joblib
python-multipart
websockets
pytest
pytest-cov
requests
passlib[bcrypt]
bcrypt
google-generativeai
kagglehub
email-validator
torch
sentence-transformers
scipy

# Blockchain dependencies
web3==6.15.0
eth-account==0.10.0
